+++ src/encord_active/lib/metrics/geometric/hu_static.py
@@ -8,13 +8,13 @@ from encord_active.lib.common.iterator import Iterator
 from encord_active.lib.common.utils import get_object_coordinates
 from encord_active.lib.embeddings.hu_moments import get_hu_embeddings
 from encord_active.lib.embeddings.writer import CSVEmbeddingWriter
-from encord_active.lib.metrcis.writer import CSVMetricWriter
 from encord_active.lib.metrics.metric import (
     AnnotationType,
     DataType,
     Metric,
     MetricType,
 )
+from encord_active.lib.metrics.writer import CSVMetricWriter
 
 logger = logger.opt(colors=True)
