+++ src/encord_active/lib/metrics/semantic/_class_uncertainty.py
@@ -10,7 +10,7 @@ from torch import nn
 from torch.nn import LeakyReLU
 
 from encord_active.lib.common.iterator import Iterator
-from encord_active.lib.embeddings.cnn_embed import get_cnn_embeddings
+from encord_active.lib.embeddings.cnn import get_cnn_embeddings
 from encord_active.lib.metrics.metric import (
     AnnotationType,
     DataType,
