+++ src/encord_active/lib/metrics/semantic/img_classification_quality.py
@@ -10,7 +10,7 @@ from encord.objects.common import PropertyType
 from loguru import logger
 
 from encord_active.lib.common.iterator import Iterator
-from encord_active.lib.embeddings.cnn_embed import get_cnn_embeddings
+from encord_active.lib.embeddings.cnn import get_cnn_embeddings
 from encord_active.lib.metrics.metric import (
     AnnotationType,
     DataType,
